shared: fcntl uses int, not long
fcntl(2) manual says the return type is int, and that F_SETFD takes an
int. So use int.

Noticed by code inspection.

+++ shared/os-compatibility.c
@@ -43,7 +43,7 @@
 int
 os_fd_set_cloexec(int fd)
 {
-	long flags;
+	int flags;
 
 	if (fd == -1)
 		return -1;
